moved FE XL into frontend, fixed eslint errors
