+++ frontend/.eslintrc.json
@@ -37,5 +37,6 @@
       }
     ],
     "react/display-name": 1
+    // "react/prop-types": "off"
   }
 }
