+++ frontend/src/Components/Navbars/Navbar.css
@@ -1,7 +1,7 @@
 
   .nav-text {
     display: flex;
-    justify-content: start;
+    justify-content: flex-start;
     align-items: center;
     /* padding: 8px 8px 8px 16px; */
     list-style: none;
