+++ frontend/src/Components/Navbars/Navbar.scss
